\name{cmscu-package}
\alias{cmscu-package}
\alias{cmscu}
\docType{package}
\title{A Count-Min-Sketch for R.}
\description{For use in obtaining frequency data of ngrams in a large corpus of textual data.}
\details{
\tabular{ll}{
Package: \tab cmscu\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-12-31\cr
License: \tab UNLICENSE\cr
}
There are only 4 methods: store(string), parallel_store(num threads, vector of strings), query(string), parallel_query(num threads, vector of strings).
}
\author{
Jason K Davis
Maintainer: Jason K Davis <jkd@jasonkarldavis.com>
}
\references{http://www.sciencedirect.com/science/article/pii/S0196677403001913}
\keyword{ ngram, sketch }
\seealso{}
\examples{}

